pyproject.toml
[project]
name = "adk-dynamodb-session"
version = "0.0.3"
description = "Session Service with DynamoDB as the backend for Google ADK"
readme = "README.md"
authors = [
    { name = "Sachdeva, Kapil"}
]
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.38",
    "google-adk>=1.5.0",
    "httpx>=0.28.1",
    "pynamodb>=6.1.0",
    "types-boto3[essential]>=1.38.38",
]

classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
]

[project.urls]
repository = "https://github.com/ksachdeva/adk-dynamodb-session"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.6.1",
    "pytest-mock>=3.14.0",
    "poethepoet>=0.30.0",
    "packaging>=24.2",
    "pyright>=1.1.378",
    "mypy>=1.10.0",
    "ruff>=0.4.8",
    "litellm>=1.72.6.post1",
    "pytest-env>=1.1.5",
]

[tool.ruff]
line-length = 120
fix = true
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501"]

[tool.ruff.lint.flake8-tidy-imports]
[tool.ruff.lint.flake8-tidy-imports.banned-api]
"unittest".msg = "Use `pytest` instead."

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false

disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false

[tool.pyright]
include = ["src", "tests", "samples"]
# typeCheckingMode = "strict"
reportUnnecessaryIsInstance = false
reportMissingTypeStubs = false

[tool.pytest.ini_options]
env = [
 "AWS_ACCESS_KEY_ID=1", 
 "AWS_SECRET_ACCESS_KEY=2", 
 "AWS_DEFAULT_REGION=us-east-1",
 "AWS_ENDPOINT_URL_DYNAMODB=http://host.docker.internal:8009" 
]

[tool.poe.tasks]
test = "pytest"
